(PC-25038) refactor(BlockDescription): Reuse containers for underage and eighteen descriptions
