+++ src/features/tutorial/components/profileTutorial/UnderageBlockDescription.tsx
@@ -4,10 +4,11 @@ import styled from 'styled-components/native'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { CreditProgressBar } from 'features/profile/components/CreditInfo/CreditProgressBar'
 import { isUserUnderageBeneficiary } from 'features/profile/helpers/isUserUnderageBeneficiary'
+import { BlockDescriptionItem } from 'features/tutorial/components/profileTutorial/BlockDescriptionItem'
 import { AccessibilityList } from 'ui/components/accessibility/AccessibilityList'
 import { BicolorClock } from 'ui/svg/icons/BicolorClock'
 import { BicolorLock } from 'ui/svg/icons/BicolorLock'
-import { Spacer, Typo, getSpacing } from 'ui/theme'
+import { Spacer } from 'ui/theme'
 
 export const UnderageBlockDescription: () => React.ReactElement = () => {
   const { isLoggedIn, user } = useAuthContext()
@@ -22,21 +23,6 @@ export const UnderageBlockDescription: () => React.ReactElement = () => {
   )
 }
 
-const ItemContainer = styled.View({
-  flexDirection: 'row',
-  alignItems: 'center',
-})
-
-const StyledCaption = styled(Typo.Caption)({
-  flexShrink: 1,
-})
-
-const IconContainer = styled.View(({ theme }) => ({
-  marginRight: getSpacing(2),
-  width: theme.icons.sizes.extraSmall,
-  height: theme.icons.sizes.extraSmall,
-}))
-
 const SmallLock = styled(BicolorLock).attrs(({ theme }) => ({
   size: theme.icons.sizes.extraSmall,
 }))``
@@ -46,18 +32,10 @@ const SmallClock = styled(BicolorClock).attrs(({ theme }) => ({
 }))``
 
 const defaultItems = [
-  <ItemContainer key={1}>
-    <IconContainer>
-      <SmallLock />
-    </IconContainer>
-    <StyledCaption>Tu as 1 an pour activer ton crédit.</StyledCaption>
-  </ItemContainer>,
-  <ItemContainer key={2}>
-    <IconContainer>
-      <SmallClock />
-    </IconContainer>
-    <StyledCaption>
-      Après activation, tu peux dépenser ton crédit jusqu’à la veille de tes 18 ans.
-    </StyledCaption>
-  </ItemContainer>,
+  <BlockDescriptionItem key={1} icon={<SmallLock />} text="Tu as 1 an pour activer ton crédit." />,
+  <BlockDescriptionItem
+    key={2}
+    icon={<SmallClock />}
+    text="Après activation, tu peux dépenser ton crédit jusqu’à la veille de tes 18 ans."
+  />,
 ]
